move UI state to the UI component
