+++ packages/app/app/containers/SearchBoxContainer/index.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import _ from 'lodash';
+import React, { useCallback } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { useTranslation } from 'react-i18next';
@@ -47,25 +46,6 @@ const SearchBoxContainer: React.FC = () => {
     provider && dispatch(selectMetaProvider(provider.value)),
   [dispatch]);
 
-  const [input, setInput] = useState('');
- 
-  const debouncedSearch = useCallback(_.debounce(handleSearch, 500), [handleSearch]);
-
-  useEffect(() => {
-    if (input.length > MIN_SEARCH_LENGTH) {
-      debouncedSearch(input);
-    }
-  }, [input, debouncedSearch]);
-
-  
-  const onKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleSearch(input);
-    }
-    if (e.key === 'Escape') {
-      handleFocus(false);
-    }
-  };
 
   return <SearchBox 
     loading={unifiedSearchStarted}
@@ -82,9 +62,7 @@ const SearchBoxContainer: React.FC = () => {
     onSearchProviderSelect={handleSelectSearchProvider}
     isFocused={isFocused}
     handleFocus={handleFocus}
-    onKeyDown={onKeyDown}
-    input={input}
-    setInput={setInput}
+    minSearchLength={MIN_SEARCH_LENGTH}
   />;
 };
